@@ -37,54 +37,53 @@
     }
   },
   "dependencies": {
-    "@godaddy/terminus": "^4.10.2",
+    "@godaddy/terminus": "^4.12.1",
     "@map-colonies/error-express-handler": "^2.1.0",
     "@map-colonies/express-access-log-middleware": "^1.0.0",
     "@map-colonies/js-logger": "^0.0.5",
     "@map-colonies/node-osm-elements": "^2.1.2",
     "@map-colonies/openapi-express-viewer": "^2.0.1",
-    "@map-colonies/osm-change-generator": "^1.0.1",
+    "@map-colonies/osm-change-generator": "^1.1.0",
     "@map-colonies/telemetry": "^4.1.0",
     "@opentelemetry/api": "^1.1.0",
     "@opentelemetry/api-metrics": "^0.29.0",
     "compression": "^1.7.4",
-    "config": "^3.3.7",
-    "express": "^4.18.1",
+    "config": "^3.3.12",
+    "express": "^4.21.0",
     "express-openapi-validator": "^4.13.8",
-    "http-status-codes": "^2.2.0",
-    "reflect-metadata": "^0.1.13",
-    "tsyringe": "^4.6.0"
+    "http-status-codes": "^2.3.0",
+    "reflect-metadata": "^0.2.2",
+    "tsyringe": "^4.8.0"
   },
   "devDependencies": {
-    "@commitlint/cli": "^17.0.2",
-    "@commitlint/config-conventional": "^17.0.2",
-    "@map-colonies/eslint-config": "^3.1.1",
+    "@commitlint/cli": "^19.5.0",
+    "@commitlint/config-conventional": "^19.5.0",
+    "@map-colonies/eslint-config": "^4.0.0",
     "@map-colonies/prettier-config": "0.0.1",
     "@map-colonies/standard-version-update-helm-version": "^2.0.1",
     "@map-colonies/standard-version-update-openapi3-version": "0.0.2",
-    "@redocly/openapi-cli": "^1.0.0-beta.94",
-    "@types/compression": "^1.7.2",
-    "@types/config": "0.0.41",
+    "@redocly/openapi-cli": "^1.0.0-beta.95",
+    "@types/compression": "^1.7.5",
+    "@types/config": "^3.3.5",
     "@types/express": "^4.17.13",
-    "@types/geojson": "^7946.0.8",
-    "@types/jest": "^28.1.0",
-    "@types/multer": "^1.4.7",
+    "@types/geojson": "^7946.0.14",
+    "@types/jest": "^29.5.13",
+    "@types/multer": "^1.4.12",
     "@types/supertest": "^2.0.12",
-    "@types/yamljs": "^0.2.31",
-    "commitlint": "^17.0.2",
+    "@types/yamljs": "^0.2.34",
+    "commitlint": "^19.5.0",
     "copyfiles": "^2.4.1",
     "cz-conventional-changelog": "^3.3.0",
     "eslint": "^8.16.0",
     "husky": "^8.0.1",
-    "jest": "^28.1.0",
+    "jest": "^29.7.0",
     "jest-create-mock-instance": "^2.0.0",
-    "jest-html-reporters": "^3.0.8",
-    "nodemon": "^2.0.16",
-    "prettier": "^2.6.2",
+    "jest-html-reporters": "^3.1.7",
+    "prettier": "^3.3.3",
     "rimraf": "^3.0.2",
     "standard-version": "^9.5.0",
     "supertest": "^6.2.3",
-    "ts-jest": "^28.0.3",
-    "typescript": "^4.6.3"
+    "ts-jest": "^29.2.5",
+    "typescript": "^5.6.2"
   }
 }
